ref&style: adjust calendar classes
