+++ shared/src/components/timetable/calendar/index.tsx
@@ -21,20 +21,21 @@ import {
   calculateDiv,
   calculateMinifiedDiv,
   calendarId,
+  CLASS_DIV_CONTAINER_ID,
+  CLASS_DIV_ID,
   classTile,
   classWeekday,
   divActiveId,
   divNowId,
   getWeek,
   minificationFrameTime,
+  placeCalendarClassDivs,
   Value,
   weekdaysEqual,
 } from './calendarUtils';
 import { Timetable } from '../../../network/models/psu-tools/timetable';
 
 const CALENDAR_MINIFIED_KEY = 'calendar-minified';
-const CLASS_DIV_ID = 'class-div';
-const CLASS_DIV_CONTAINER_ID = 'class-div-container';
 
 const calculateNowDivMinified = () =>
   calculateMinifiedDiv(divNowId, `${classWeekday}--now-minified`); // prettier-ignore
@@ -133,48 +134,7 @@ const CustomCalendar = ({
   useEffect(() => {
     if (!classesData) return;
 
-    const curMonth = activeStartDate.getMonth();
-    const curYear = activeStartDate.getFullYear();
-    const curMonthString = `${(curMonth + 1).toString().padStart(2, '0')}.${curYear}`;
-
-    const classDiv = node(`#${CLASS_DIV_ID}`);
-    const classDivContainer = node(`#${CLASS_DIV_CONTAINER_ID}`);
-    let firstDayOfCurMonthIndex: number;
-
-    const tiles = nodes(`.react-calendar__month-view__days__day`);
-
-    for (let i = 0; i < tiles.length; i++) {
-      if (
-        tiles[i].classList.contains(
-          'react-calendar__month-view__days__day--neighboringMonth',
-        )
-      )
-        // eslint-disable-next-line no-continue
-        continue;
-
-      firstDayOfCurMonthIndex = i - 1;
-      break;
-    }
-
-    classesData.forEach((day) => {
-      if (!day.date.endsWith(curMonthString)) return;
-
-      const tile =
-        tiles[firstDayOfCurMonthIndex + parseInt(day.date.substring(0, 2))];
-
-      tile.classList.add('relative');
-      const containerCloned = classDivContainer?.cloneNode() as Element;
-      containerCloned.classList.add('flex');
-      containerCloned.classList.remove('hidden');
-
-      day.classes.forEach(() => {
-        const cloned = classDiv?.cloneNode() as Element;
-        cloned.classList.remove('hidden');
-        containerCloned.appendChild(cloned);
-      });
-
-      tile.appendChild(containerCloned);
-    });
+    placeCalendarClassDivs(activeStartDate, classesData);
   }, [classesData, value, isMinified, activeStartDate]);
 
   const tileClassName = ({ date }: { date: Date }) => {
@@ -262,11 +222,11 @@ const CustomCalendar = ({
     <div className={`${className}`}>
       <div
         id={CLASS_DIV_ID}
-        className="size-[0.25rem] rounded-sm bg-white hidden left-0"
+        className="size-[0.25rem] rounded-sm hidden left-0 border-0"
       />
       <div
         id={CLASS_DIV_CONTAINER_ID}
-        className="absolute bottom-0 size-fit flex-row gap-1 left-1/2 -translate-x-1/2 hidden"
+        className="absolute bottom-1 size-fit flex-row gap-[0.15rem] left-1/2 -translate-x-1/2 hidden"
       />
       <Button
         type="button"
